tests: replace superlock test with unified spinlock
